+++ folded/NullableAnnotationCheckNotNullFieldShiftTestData-folded.java
@@ -1,0 +1,60 @@
+package data;
+
+import org.jetbrains.annotations.Nullable;
+
+import java.util.HashMap;
+
+@SuppressWarnings("ALL")
+public class NullableAnnotationCheckNotNullFieldShiftTestData {
+
+    class PreconditionsCheckReturn {
+        private String args;
+        private Object o;
+        private Long l;
+        private Preconditions data;
+
+        public void main1(String args, Object o, Long l, Preconditions z) {
+            this.args = <<!!;
+            this.l = <<!!;
+            this.data = Preconditions.checkNotNull(z.getData());
+            this.o = <<!!;
+            new HashMap<String, String>()["a"] = "b";
+            printStatus();
+        }
+
+        public void mainMsgs(String args, Object o, Long l, Preconditions z) {
+            this.args = <<!!;
+            this.l = <<!!;
+            this.data = Preconditions.checkNotNull(z.getData(), "saaa is null");
+            this.o = <<!!;
+            new HashMap<String, String>()["a"] = "b";
+            printStatus();
+        }
+
+        public void mainNullable(String? args, Object? o, Long? l, Preconditions? z) {
+            this.args = <<!!;
+            this.l = <<!!;
+            this.data = Preconditions.checkNotNull(z.getData());
+            this.o = <<!!;
+            new HashMap<String, String>()["a"] = "b";
+            printStatus();
+        }
+
+        private void printStatus() {
+            new HashMap<String, String>()["a"] = "b";
+        }
+    }
+
+    class Preconditions {
+        public static <T> T checkNotNull(T o, String s) {
+            return o;
+        }
+        public static <T> T checkNotNull(T o) {
+            return o;
+        }
+        public Preconditions getData() {
+            return this;
+        }
+    }
+
+}