+++ testData/NullableAnnotationCheckNotNullTestData-all.java
@@ -27,7 +27,7 @@ public class NullableAnnotationCheckNotNullTestData {
         }</fold>
 
         public void mainConflictAnnotations(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
-            <fold text='' expand='false'><fold text='' expand='true'></fold>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;</fold><fold text='' expand='true'>
+            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;<fold text='' expand='true'></fold>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>)</fold>;</fold>
             <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>)</fold>;
             <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;
@@ -35,7 +35,7 @@ public class NullableAnnotationCheckNotNullTestData {
             new HashMap<String, String>()<fold text='[' expand='false'>.put(</fold>"a"<fold text='] = ' expand='false'>, </fold>"b"<fold text='' expand='false'>)</fold>;
         }</fold>
 
-        public void mainConflictAnnotationsWithMsg(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
+        public void mainConflictAnnotationsWithMsg(<fold text='' expand='false'>@Nullable<fold text='' expand='false'></fold> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;</fold><fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;</fold>
             <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>, "o is null")</fold>;
@@ -61,10 +61,10 @@ public class NullableAnnotationCheckNotNullTestData {
         }</fold>
 
         public void mainMsgs(String args, Object o, Long l, Preconditions z) <fold text='{...}' expand='true'>{
-            this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;
-            this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;
-            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>, "saaa is null")</fold>;
-            this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;
+            this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold><fold text='<<' expand='false'>args</fold><fold text='!!' expand='false'>, "args are null")</fold>;
+            this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold><fold text='<<' expand='false'>l</fold><fold text='!!' expand='false'>, "l is null")</fold>;
+            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='<<' expand='false'>getData()</fold><fold text='!!' expand='false'>, "saaa is null")</fold>;
+            this.o = <fold text='' expand='false'>Preconditions.checkNotNull(<fold text='<<' expand='false'></fold>o</fold><fold text='!!' expand='false'>, "o is null")</fold>;
             new HashMap<String, String>()<fold text='[' expand='false'>.put(</fold>"a"<fold text='] = ' expand='false'>, </fold>"b"<fold text='' expand='false'>)</fold>;
             printStatus();
         }</fold>
@@ -79,10 +79,10 @@ public class NullableAnnotationCheckNotNullTestData {
         }</fold>
 
         public void mainMsgsNullable(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='? ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='? ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{
-            this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;
-            this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;
-            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>, "saaa is null")</fold>;
-            this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;
+            this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold><fold text='<<' expand='false'>args</fold><fold text='!!' expand='false'>, "args are null")</fold>;
+            this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold><fold text='<<' expand='false'>l</fold><fold text='!!' expand='false'>, "l is null")</fold>;
+            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='<<' expand='false'>getData()</fold><fold text='!!' expand='false'>, "saaa is null")</fold>;
+            this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold><fold text='<<' expand='false'>o</fold><fold text='!!' expand='false'>, "o is null")</fold>;
             new HashMap<String, String>()<fold text='[' expand='false'>.put(</fold>"a"<fold text='] = ' expand='false'>, </fold>"b"<fold text='' expand='false'>)</fold>;
             printStatus();
         }</fold>