+++ testData/NullableAnnotationCheckNotNullTestData.java
@@ -8,38 +8,38 @@ import java.util.HashMap;</fold>
 public class NullableAnnotationCheckNotNullTestData {
 
     class PreconditionsCheck <fold text='{...}' expand='true'>{
-        public void main(String<fold text='!!! ' expand='false'> </fold>args, Object<fold text='!!! ' expand='false'> </fold>o, Long<fold text='!!! ' expand='false'> </fold>l, Preconditions z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
-            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;<fold text='' expand='true'></fold>
-            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>)</fold>;</fold>
-            Preconditions.checkNotNull(z.getData());<fold text='' expand='true'>
-            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;</fold>
+        public void main(String<fold text='!!! ' expand='false'> </fold>args, Object o, Long<fold text='!!! ' expand='false'> </fold>l, Preconditions z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
+            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;</fold><fold text='' expand='true'>
+            <fold text='' expand='false'><fold text='' expand='true'></fold>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>)</fold>;</fold>
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>)</fold>;
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;
             System.out.println();
             new HashMap<String, String>().put("a", "b");
         }</fold>
 
-        public void mainMsgs(String<fold text='!!! ' expand='false'> </fold>args, Object<fold text='!!! ' expand='false'> </fold>o, Long<fold text='!!! ' expand='false'> </fold>l, Preconditions z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
+        public void mainMsgs(String<fold text='!!! ' expand='false'> </fold>args, Object o, Long<fold text='!!! ' expand='false'> </fold>l, Preconditions z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;</fold><fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;</fold>
-            Preconditions.checkNotNull(z.getData(), "o is null");<fold text='' expand='true'>
-            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;</fold>
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>, "o is null")</fold>;
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;
             System.out.println();
             new HashMap<String, String>().put("a", "b");
         }</fold>
 
-        public void mainConflictAnnotations(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable<fold text='' expand='false'></fold> </fold>Object<fold text='!!! ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
+        public void mainConflictAnnotations(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;</fold><fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>)</fold>;</fold>
-            Preconditions.checkNotNull(z.getData());<fold text='' expand='true'>
-            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;</fold>
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()<fold text='!!' expand='false'></fold>)</fold>;
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;
             System.out.println();
             new HashMap<String, String>().put("a", "b");
         }</fold>
 
-        public void mainConflictAnnotationsWithMsg(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='!!! ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
+        public void mainConflictAnnotationsWithMsg(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='!!! ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='!!! ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{<fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;</fold><fold text='' expand='true'>
             </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;</fold>
-            Preconditions.checkNotNull(z.getData(), "o is null");<fold text='' expand='true'>
-            </fold><fold text='' expand='false'><fold text='' expand='true'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;</fold>
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()<fold text='!!' expand='false'></fold>, "o is null")</fold>;
+            <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;
             new HashMap<String, String>().put("a", "b");
             System.out.println();
         }</fold>
@@ -54,7 +54,7 @@ public class NullableAnnotationCheckNotNullTestData {
         public void main1(String args, Object o, Long l, Preconditions z) <fold text='{...}' expand='true'>{
             this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;
             this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>)</fold>;
-            this.data = Preconditions.checkNotNull(z.getData());
+            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>)</fold>;
             this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;
             new HashMap<String, String>().put("a", "b");
             printStatus();
@@ -63,7 +63,7 @@ public class NullableAnnotationCheckNotNullTestData {
         public void mainMsgs(String args, Object o, Long l, Preconditions z) <fold text='{...}' expand='true'>{
             this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;
             this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;
-            this.data = Preconditions.checkNotNull(z.getData(), "saaa is null");
+            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>, "saaa is null")</fold>;
             this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;
             new HashMap<String, String>().put("a", "b");
             printStatus();
@@ -72,7 +72,7 @@ public class NullableAnnotationCheckNotNullTestData {
         public void mainNullable(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='? ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='? ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{
             this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>)</fold>;
             this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>)</fold>;
-            this.data = Preconditions.checkNotNull(z.getData());
+            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>)</fold>;
             this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>)</fold>;
             new HashMap<String, String>().put("a", "b");
             printStatus();
@@ -81,7 +81,7 @@ public class NullableAnnotationCheckNotNullTestData {
         public void mainMsgsNullable(<fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>String<fold text='? ' expand='false'> </fold>args, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Object<fold text='? ' expand='false'> </fold>o, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Long<fold text='? ' expand='false'> </fold>l, <fold text='' expand='false'>@Nullable</fold><fold text='' expand='false'> </fold>Preconditions<fold text='? ' expand='false'> </fold>z) <fold text='{...}' expand='true'>{
             this.args = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>args<fold text='!!' expand='false'>, "args are null")</fold>;
             this.l = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>l<fold text='!!' expand='false'>, "l is null")</fold>;
-            this.data = Preconditions.checkNotNull(z.getData(), "saaa is null");
+            this.data = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>z.<fold text='data' expand='false'>getData()</fold><fold text='!!' expand='false'>, "saaa is null")</fold>;
             this.o = <fold text='' expand='false'>Preconditions.checkNotNull(</fold>o<fold text='!!' expand='false'>, "o is null")</fold>;
             new HashMap<String, String>().put("a", "b");
             printStatus();